+++ Integrations/ProofpointTAP_v2/CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [Unreleased]
+The **fetch-incidents** function does not fetch duplicate values.
+
+## [19.10.2] - 2019-10-29
 Fixed the **fetch-incidents** function when the last_fetch time range is greater than 1 hour.
 
 ## [19.8.0] - 2019-08-06